refactor(web): extract Plan cluster to pages/runView/PlanPanel.tsx (#1007 M1)
Pure motion: MAX_REVISION_ROUNDS, PlanRevision, derivePlanRevision,
VersionChip, RevisionThread, PlanPanel, SeededPlanPanel move verbatim.
RunView.tsx re-exports PlanPanel/SeededPlanPanel/derivePlanRevision so
importers and tests stay byte-identical. specs/ai.md location repointed.
